Update NYS.Rmd
Use Loops to Reduce Repetition
The code is repetitive, doing essentially the same thing for each package: checking if it is installed, installing it if it is not, and then loading it. You can use loops to handle this in a more concise and maintainable way.

Use requireNamespace instead of require
The require function attaches the package, which can sometimes have unintended side effects. requireNamespace only checks if the package is installed without attaching it.

Error Handling
add error handling, for example, by using tryCatch around install.packages, to handle scenarios where a package fails to install.

Enable Parallel Installation
enable parallel installation of packages using Ncpus parameter in install.packages.