Update NYS.Rmd
encapsulated the entire process into a function called download_and_process_data(). This function tries to download the data, and if it succeeds, it processes the data and returns a cleaned dataset. I've also streamlined the code to calculate incremental cases (I) directly in the same pipeline where the data is downloaded and cleaned.